# internal/api/swagger/swagger.yaml
basePath: /
consumes:
- application/json
host: localhost:8080
info:
  contact:
    name: Nicholas Fedor
    url: https://github.com/nicholas-fedor/watchtower
  description: Watchtower HTTP API for container update management, metrics, and health
    probes.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Watchtower HTTP API
  version: "1.0"
paths:
...
  /v1/containers/details:
    get:
      consumes:
      - application/json
      description: Returns detailed information about each watched container, including
        running state, image identity, and configuration flags. Optionally filter
        by container name or image name.
      parameters:
      - description: Filter by container name (exact match)
        in: query
        name: name
        type: string
      - description: Filter by image name (exact match)
        in: query
        name: image
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Container details with count and timestamp
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid API token
          schema:
            type: string
        "500":
          description: Failed to get container details
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Detailed container status
      tags:
      - containers
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Paste the HTTP API token only (WATCHTOWER_HTTP_API_TOKEN). Swagger
      UI sends it as the Authorization header value. curl clients should use "Authorization:
      Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
  EventsToken:
    description: Paste the events API token only (WATCHTOWER_HTTP_API_EVENTS_TOKEN).
      Swagger UI sends it as the Authorization header value.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"